Load onionshare_cli from source tree instead of pip dependency, and start making tests work with PySide2
