+++ desktop/README.md
@@ -58,4 +58,10 @@ Then run the tests:
 ./tests/run.sh
 ```
 
+If you want to run tests while hiding the GUI, you must have the `xorg-x11-server-Xvfb` package installed, and then:
+
+```sh
+xvfb-run ./tests/run.sh
+```
+
 ## Making a release
