+++ desktop/tests/run.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-pytest -vvv --no-qt-log tests/test_gui_tabs.py
-pytest -vvv --no-qt-log tests/test_gui_share.py
-pytest -vvv --no-qt-log tests/test_gui_receive.py
-pytest -vvv --no-qt-log tests/test_gui_website.py
+pytest -v tests/test_gui_tabs.py
+pytest -v tests/test_gui_share.py
+pytest -v tests/test_gui_receive.py
+pytest -v tests/test_gui_website.py
